text_editor/requirements.txt
tkinter==8.6
pillow==10.1.0
